add a player class and player tests
this commit adds three methods to the player class, cpuGenAttack(),
cpuTurn() and playerTurn().

-cpuGenAttack()
generates a random integer between 0-9 for both the x and y
coordinates. If the space has already been attacked (boardArray is
noted by either 'M' or 'H') another random coordinate will be
generated.

-cpuTurn()
will generate a coordinate cpuGenAttack() and then attack the player
gameboard with those generated coordinates.

-playerTurn()
attacks the cpu gameboard with an attack parameter passed to this
method.

tests were also implimented to test that cpuGenAttack generates
a valid coordinate and that the playerTurn registers an attack
properly on the CpuBoard.