allow randomized placement of ships on load
In an effort to try to keep this game entertaining this commit creates
a method of the Gameboard class called genShipCoords() which takes
a ship length as an argument and returns valid ship coordinates.
This method has checks that will automatically verify by referencing
boardArray that the space is valid and a ship is placeable there.
A test was also added to verify that the genShipCoords() method will
return an array with the length of the ship length passed into the
method initally.