improve CPU AI within cpuGenAttack() method
This commit improves the computer AI by storing the last
successful hit in the Player class property lastCpuAttack and
attempting to attack adjacent squares based on this information.
I did not write any additional tests for this as I figure there is
some randomness involved and the only thing we can really be sure of is
that we will recieve coordinates in an array and that array length sould
always be two.